Bunch of updates with droptable, new mechanics & support for more things, like crystal chest, better collection log handling

+++ src/main/java/com/rs/java/game/objects/plugins/CrystalChestObject.java
@@ -4,11 +4,17 @@ import com.rs.java.game.Animation;
 import com.rs.java.game.World;
 import com.rs.java.game.WorldObject;
 import com.rs.java.game.item.Item;
+import com.rs.java.game.item.ground.GroundItems;
 import com.rs.java.game.objects.ObjectPlugin;
 import com.rs.java.game.player.Player;
 import com.rs.java.game.player.content.CrystalChest;
 import com.rs.core.tasks.WorldTask;
 import com.rs.core.tasks.WorldTasksManager;
+import com.rs.kotlin.game.npc.drops.Drop;
+import com.rs.kotlin.game.npc.drops.tables.CrystalChestTable;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 public class CrystalChestObject extends ObjectPlugin {
 
@@ -38,13 +44,25 @@ public class CrystalChestObject extends ObjectPlugin {
 	}
 
 	public void openChest(Player player, WorldObject object) {
+		if (!player.getInventory().hasFreeSlots()) {
+			player.message("You don't have any space in your inventory.");
+			return;
+		}
 		player.animate(new Animation(536));
 		player.lock(2);
 		player.getPackets().sendGameMessage("You attempt to unlock the chest...");
 		WorldObject openedChest = new WorldObject(173, object.getType(), object.getRotation(), object);
 		World.replaceObjectTemporary(object, openedChest, 1, () -> {
 			player.getInventory().deleteItem(CRYSTAL_KEY, 1);
-			CrystalChest.sendRewards(false, player);
+			player.getKillcount().increment("Crystal chest");
+			List<Drop> drops = CrystalChestTable.INSTANCE.getTable().rollDrops(player, 0);
+			for (Drop drop : drops) {
+				if (!player.inventory.canHold(drop.itemId, drop.amount)) {
+					GroundItems.updateGroundItem(new Item(drop.itemId, drop.amount), player.getLocation(), player, 60);
+					continue;
+				}
+				player.getInventory().addItem(drop.itemId, drop.amount);
+			}
 		});
 	}
 }
